Shader graph effects, wave spawner 2.0

+++ TowerDefense/Assets/Scripts/Bullet.cs
@@ -4,11 +4,10 @@ public class Bullet : MonoBehaviour
 {
     Transform target;
     public float speed = 50f;
-    public int damage = 50;
+    public float damage = 50f;
     public float explosionRadius = 0f;
     public GameObject impactEffect;
 
-    // Update is called once per frame
     void Update()
     {
         if(target == null)
